Move run_server from conftest to utils
